@@ -5,21 +5,21 @@ import 'package:flutter_auto_form/src/models/form.dart';
 import 'package:flutter_auto_form/src/widgets/form_state.dart';
 
 class AutoFormWidget<T extends TemplateForm> extends StatefulWidget {
-  const AutoFormWidget(
-      {Key key,
-      @required this.formBuilder,
-      @required this.onSubmitted,
-      this.submitButton,
-      this.handleErrorOnSubmit})
-      : super(key: key);
+  const AutoFormWidget({
+    Key? key,
+    required this.formBuilder,
+    required this.onSubmitted,
+    this.submitButton,
+    this.handleErrorOnSubmit,
+  }) : super(key: key);
 
   final T Function() formBuilder;
 
   final Function(T form) onSubmitted;
 
-  final ValueChanged<String> handleErrorOnSubmit;
+  final ValueChanged<String>? handleErrorOnSubmit;
 
-  final Widget Function(Function(bool showLoadingDialog) submit) submitButton;
+  final Widget Function(Function(bool showLoadingDialog) submit)? submitButton;
 
   @override
   _AutoFormWidgetState createState() => _AutoFormWidgetState<T>(
@@ -30,7 +30,7 @@ class AutoFormWidget<T extends TemplateForm> extends StatefulWidget {
 
 class _AutoFormWidgetState<T extends TemplateForm>
     extends AutoFormWidgetState<AutoFormWidget, T> {
-  _AutoFormWidgetState(T model, {ValueChanged<String> handleErrorOnSubmit})
+  _AutoFormWidgetState(T model, {ValueChanged<String>? handleErrorOnSubmit})
       : super(model: model, handleErrorOnSubmit: handleErrorOnSubmit);
 
   @override
@@ -41,7 +41,7 @@ class _AutoFormWidgetState<T extends TemplateForm>
       child = Column(
         children: [
           child,
-          widget.submitButton(
+          widget.submitButton!(
             (bool showLoadingDialog) => submitForm(
               showLoading: showLoadingDialog,
             ),
