@@ -6,7 +6,7 @@ class URLValidator extends ValidatorWithStaticError<String> {
   URLValidator(String error) : super(error);
 
   @override
-  bool innerValidate(String value) {
+  bool innerValidate(String? value) {
     return value == null || RegExp(emailRegex).allMatches(value).isNotEmpty;
   }
 }
@@ -15,7 +15,7 @@ class HexColorValidator extends ValidatorWithStaticError<String> {
   HexColorValidator(String error) : super(error);
 
   @override
-  bool innerValidate(String value) {
+  bool innerValidate(String? value) {
     return value != null &&
         value.length == 6 &&
         RegExp(hexRegex).hasMatch(value);
@@ -26,18 +26,18 @@ class NotNullValidator extends ValidatorWithStaticError<String> {
   NotNullValidator(String error) : super(error);
 
   @override
-  bool innerValidate(String value) => value != null;
+  bool innerValidate(String? value) => value != null;
 }
 
 class MinimumStringLengthValidator extends Validator<String> {
   MinimumStringLengthValidator(this.minStringLength, this.error);
 
-  final String Function(String value) error;
+  final String Function(String? value) error;
 
   final int minStringLength;
 
   @override
-  String validate(String value) {
+  String? validate(String? value) {
     if (value == null || value.length < minStringLength) return error(value);
 
     return null;
@@ -50,14 +50,14 @@ class SameAsFieldValidator<T> extends ValidatorWithStaticError<T> {
   final Field field;
 
   @override
-  bool innerValidate(T value) => value == field.value;
+  bool innerValidate(T? value) => value == field.value;
 }
 
 class AlphanumericValidator extends ValidatorWithStaticError<String> {
   AlphanumericValidator(String error) : super(error);
 
   @override
-  bool innerValidate(String value) =>
+  bool innerValidate(String? value) =>
       value != null && RegExp(alphaNumericRegex).hasMatch(value);
 }
 
@@ -65,9 +65,7 @@ class EmailValidator extends ValidatorWithStaticError<String> {
   EmailValidator(String error) : super(error);
 
   @override
-  bool innerValidate(String value) {
-    print(value);
-    print(RegExp(emailRegex).hasMatch(value));
+  bool innerValidate(String? value) {
     return value != null && RegExp(emailRegex).hasMatch(value);
   }
 }
